requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
whitenoise
psycopg2-binary
drf-yasg
